2024-06-02 typos

+++ 1064/submissions/cake.py
@@ -1,11 +1,11 @@
 layer1_height = float(input("Enter height of layer 1 (cm): "))
 layer1_diameter = float(input("Enter diameter of layer 1 (cm): "))
 
-layer2_height = float(input("Enter height of layer 2 (cm): "))  
-layer2_diameter = float(input("Enter diameter of layer 2 (cm): "))
+layer2_height = int(input("Enter height of layer 2 (cm): "))  
+layer2_diameter = int(input("Enter diameter of layer 2 (cm): "))
 
-layer3_height = float(input("Enter height of layer 3 (cm): "))
-layer3_diameter = float(input("Enter diameter of layer 3 (cm): "))
+layer3_height = int(input("Enter height of layer 3 (cm): "))
+layer3_diameter = int(input("Enter diameter of layer 3 (cm): "))
 
 flavor = input("Enter cake flavor: ")
